BACKEND/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
mysql-connector-python==8.2.0
google-generativeai==0.3.2
Pillow==10.1.0
aiofiles==23.2.1
pydantic==2.5.0
langchain==0.0.350
langchain-google-genai==0.0.6
sqlalchemy==2.0.23
aiomysql==0.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
